added new method to rate a restaurant, updated the restaurant rating deserialization due to an error

+++ models/restaurant_model.py
@@ -7,4 +7,4 @@ class Restaurant(BaseModel):
     categories: list[str]
     latitude: float
     longitude: float
-    rating: list[float]
+    rating: list[int]
